+++ minishell.h
@@ -98,10 +98,9 @@ char			*find_path(char *cmdline, char **envs);
 // exec
 int 			check_builtin(char **cmdline);
 void 			exec_process(t_list *proc);
-int 			exec_cmds(t_list *cur_proc, t_cmd *cmd);
-int 			exec_pipe(t_cmd *cmd);
 int 			exec_builtin(t_cmd *cmd);
-int 			exec_execve(t_cmd *cmd);
+int 			exec_fork(t_list *cur_proc, t_cmd *cmd);
+void 			exec_child_process(t_cmd *cmd, t_cmd *next_cmd);
 
 int 			ft_env(char **envs);
 int				ft_pwd(void);
